# city-data-overload-server/flutter-api-swagger-tests.yaml
openapi: 3.0.3
info:
  title: City Data Overload - Flutter API
  description: |
    Comprehensive API for the City Data Overload Flutter mobile application.
    
    ## Architecture
    - **CoordinatorAgent System** (`/api/flutter`) - New dynamic routing system
    - **Specialized Endpoints** (`/api/v1/flutter`) - Direct agent access
    
    ## Key Features
    - 🤖 AI-powered chat responses
    - 📍 Location-based data filtering
    - 📊 Real-time event streams
    - 📱 Multimodal report submissions
    - 🎯 Predictive alerts and insights
    
    ## Authentication
    Currently using basic user ID tracking. Production should implement proper auth.
    
  version: 1.0.0
  contact:
    name: City Data Overload Team
    email: support@citydata.com
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.citydata.com
    description: Production server
...
paths:
...
  /api/v1/serpapi/fetch/all:
    post:
      tags: [Utilities]
      summary: SerpAPI Data Fetch (All Categories)
      description: |
        **✅ FIXED: No more Jackson errors!**
        
        Fetches data from all categories using SerpAPI. Handles "no results" gracefully.
        
        **Fixes Applied:**
        - JSON deserialization errors resolved
        - Graceful handling of "Google hasn't returned any results"
        - Better error logging and debugging
      parameters:
        - name: location
          in: query
          schema:
            type: string
            default: "Bengaluru, Karnataka, India"
          example: "Bengaluru, Karnataka, India"
      responses:
        '200':
          description: Data fetch completed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CityEvent'
  /api/v1/serpapi/fetch/process-store:
    post:
      tags: [Utilities]
      summary: Full AI Pipeline (SYNC - 5min timeout)
      description: |
        **✅ FIXED: Timeout and Jackson errors!**
        
        Complete pipeline: fetch + AI processing + storage.
        
        **Fixes Applied:**
        - Extended timeout from 30s to 5 minutes
        - SerpAPI error handling fixed
        - Better error responses with suggestions
        
        **⚠️ Heavy Operation:** Can take 2-5 minutes
      parameters:
        - name: location
          in: query
          schema:
            type: string
            default: "Bengaluru, Karnataka, India"
          example: "Bengaluru, Karnataka, India"
      responses:
        '200':
          description: Processing completed or error with details
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [SUCCESS, ERROR]
                  location:
                    type: string
                  totalEventsFetched:
                    type: integer
                  processedEventsStored:
                    type: integer
                  timestamp:
                    type: string
                    format: date-time
                  pipeline:
                    type: string
                    example: "COMPLETE_AI_PIPELINE"
                  error:
                    type: string
                    description: Error message if failed
                  suggestion:
                    type: string
                    description: Helpful suggestion for users
  /api/v1/serpapi/fetch/process-store/async:
    post:
      tags: [Utilities]
      summary: Full AI Pipeline (ASYNC - Returns immediately)
      description: |
        **✅ NEW: Async alternative for UI integrations!**
        
        Starts the AI processing pipeline and returns immediately with task ID.
        Processing continues in background - perfect for web UIs.
        
        **Benefits:**
        - No waiting for user
        - Immediate response
        - Background processing
        - Check logs for completion
      parameters:
        - name: location
          in: query
          schema:
            type: string
            default: "Bengaluru, Karnataka, India"
          example: "Bengaluru, Karnataka, India"
      responses:
        '202':
          description: Task started successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  taskId:
                    type: string
                    example: "task_1753554376209"
                  status:
                    type: string
                    example: "STARTED"
                  location:
                    type: string
                  message:
                    type: string
                    example: "AI processing pipeline started. Check logs for progress."
                  estimatedDuration:
                    type: string
                    example: "2-5 minutes"
                  timestamp:
                    type: string
                    format: date-time
...
components:
  schemas:
...
    CityEvent:
      type: object
      properties:
        id:
          type: string
          example: "evt_123"
        title:
          type: string
          example: "Traffic jam on MG Road"
        description:
          type: string
          example: "Heavy traffic reported due to road construction"
        category:
          type: string
          enum: [TRAFFIC, UTILITY, HEALTH, INFRASTRUCTURE, EMERGENCY, COMMUNITY]
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        location:
          type: object
          properties:
            latitude:
              type: number
              format: double
            longitude:
              type: number
              format: double
            area:
              type: string
        timestamp:
          type: string
          format: date-time
        aiSummary:
          type: string
          description: AI-generated summary
        confidenceScore:
          type: number
          format: double
          description: AI confidence (0.0-1.0)
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - {} # No authentication (for testing) 